restructured backend & added functionality to add to database
