+++ backend/.env.example
@@ -1,1 +1,3 @@
-SUPABASE_JWT_SECRET =
+SUPABASE_JWT_SECRET=
+SUPABASE_URL=
+SUPABASE_ANON_KEY=
